+++ ThirdPartyAdapters/liftoffmonetize/liftoffmonetize/src/test/java/com/google/ads/mediation/vungle/VungleRewardedAdTest.kt
@@ -162,10 +162,7 @@ class VungleRewardedAdTest {
     adapter.onAdRewarded(vungleRewardedAd)
 
     verify(rewardedAdCallback).onVideoComplete()
-    verify(rewardedAdCallback)
-      .onUserEarnedReward(
-        argThat { rewardItem -> rewardItem.type == "vungle" && rewardItem.amount == 1 }
-      )
+    verify(rewardedAdCallback).onUserEarnedReward()
     verifyNoMoreInteractions(rewardedAdCallback)
   }